update authorization for careers and volunteers

+++ HospitalProject2/Controllers/VolunteersDataController.cs
@@ -1,4 +1,6 @@
 ﻿using System;
+using System.IO;
+using System.Web;
 using System.Collections.Generic;
 using System.Data;
 using System.Data.Entity;
@@ -11,6 +13,8 @@ using System.Web.Http.Description;
 using HospitalProject2.Models;
 using System.Diagnostics;
 
+using Microsoft.AspNet.Identity;
+
 
 namespace HospitalProject2.Controllers
 {
@@ -21,9 +25,22 @@ namespace HospitalProject2.Controllers
         // GET: api/VolunteersData/ListVolunteers
         [ResponseType(typeof(VolunteersDto))]
         [HttpGet]
+        [Authorize(Roles = "Admin,Guest")]
         public IHttpActionResult ListVolunteers()
         {
-            List<Volunteers> Volunteers = db.Volunteers.ToList();
+            bool isAdmin = User.IsInRole("Admin");
+            //Admins see all, guests only see their own
+            List<Volunteers> Volunteers;
+            Debug.WriteLine("id is " + User.Identity.GetUserId());
+            if (isAdmin) Volunteers = db.Volunteers.ToList();
+            else
+            {
+                string UserId = User.Identity.GetUserId();
+                Volunteers = db.Volunteers.Where(v => v.UserID == UserId).ToList();
+            }
+
+            //List<Volunteers> Volunteers = db.Volunteers.ToList();
+
             List<VolunteersDto> VolunteersDtos = new List<VolunteersDto>();
 
             Volunteers.ForEach(v => VolunteersDtos.Add(new VolunteersDto()
@@ -34,7 +51,8 @@ namespace HospitalProject2.Controllers
                 contact = v.contact,
                 email = v.email,
                 program_id = v.program_id,
-                hours = v.hours
+                hours = v.hours,
+                UserId = v.UserID
 
             }));
       
@@ -55,7 +73,8 @@ namespace HospitalProject2.Controllers
                 l_name = v.l_name,
                 contact = v.contact,
                 email = v.email,
-                hours = v.hours
+                hours = v.hours,
+                UserId= v.UserID
 
             }));
             return Ok(VolunteersDtos);
@@ -64,6 +83,7 @@ namespace HospitalProject2.Controllers
         // GET: api/VolunteersData/FindVolunteers/5
         [ResponseType(typeof(Volunteers))]
         [HttpGet]
+        [Authorize(Roles = "Admin,Guest")]
         public IHttpActionResult FindVolunteers(int id)
         {
             Volunteers Volunteers = db.Volunteers.Find(id);
@@ -75,13 +95,18 @@ namespace HospitalProject2.Controllers
                 contact = Volunteers.contact,
                 email = Volunteers.email,
                 program_id = Volunteers.program_id,
-                hours = Volunteers.hours
+                hours = Volunteers.hours,
+                UserId = Volunteers.UserID
             };
 
             if (Volunteers == null)
             {
                 return NotFound();
             }
+            //do not process if the (user is not an admin) and (the booking does not belong to the user)
+            bool isAdmin = User.IsInRole("Admin");
+            //Forbidden() isn't a natively implemented status like BadRequest()
+            if (!isAdmin && (Volunteers.UserID != User.Identity.GetUserId())) return StatusCode(HttpStatusCode.Forbidden);
 
             return Ok(VolunteersDto);
         }
@@ -89,6 +114,7 @@ namespace HospitalProject2.Controllers
         // POST: api/VolunteersData/UpdateVolunteers/5
         [ResponseType(typeof(void))]
         [HttpPost]
+        [Authorize(Roles = "Admin,Guest")]
         public IHttpActionResult UpdateVolunteers(int id, Volunteers Volunteers)
         {
             if (!ModelState.IsValid)
@@ -101,7 +127,17 @@ namespace HospitalProject2.Controllers
                 return BadRequest();
             }
 
+            //do not process if the (user is not an admin) and (the booking does not belong to the user)
+            bool isAdmin = User.IsInRole("Admin");
+            //Forbidden() isn't a natively implemented status like BadRequest()
+            if (!isAdmin && (Volunteers.UserID != User.Identity.GetUserId()))
+            {
+                Debug.WriteLine("not allowed. booking user" + Volunteers.UserID + " user " + User.Identity.GetUserId());
+                return StatusCode(HttpStatusCode.Forbidden);
+            }
+
             db.Entry(Volunteers).State = EntityState.Modified;
+            db.Entry(Volunteers).Property(v => v.UserID).IsModified = false;
 
             try
             {
@@ -125,12 +161,15 @@ namespace HospitalProject2.Controllers
         // POST: api/VolunteersData/AddVolunteers
         [ResponseType(typeof(Volunteers))]
         [HttpPost]
+        [Authorize(Roles = "Admin,Guest")]
         public IHttpActionResult AddVolunteers(Volunteers Volunteers)
         {
             if (!ModelState.IsValid)
             {
                 return BadRequest(ModelState);
             }
+
+            Volunteers.UserID = User.Identity.GetUserId();
 
             db.Volunteers.Add(Volunteers);
             db.SaveChanges();
@@ -141,6 +180,7 @@ namespace HospitalProject2.Controllers
         // POST: api/VolunteersData/DeleteVolunteers/5
         [ResponseType(typeof(Volunteers))]
         [HttpPost]
+        [Authorize(Roles = "Admin,Guest")]
         public IHttpActionResult DeleteVolunteers(int id)
         {
             Volunteers Volunteers = db.Volunteers.Find(id);
@@ -148,6 +188,12 @@ namespace HospitalProject2.Controllers
             {
                 return NotFound();
             }
+
+            //do not process if the (user is not an admin) and (the booking does not belong to the user)
+            bool isAdmin = User.IsInRole("Admin");
+            //Forbidden() isn't a natively implemented status like BadRequest()
+            if (!isAdmin && (Volunteers.UserID != User.Identity.GetUserId())) return StatusCode(HttpStatusCode.Forbidden);
+
 
             db.Volunteers.Remove(Volunteers);
             db.SaveChanges();
